refactor: controller에 연결단만 존재하도록 분리
