+++ src/main/java/com/example/be/auth/AuthController.java
@@ -3,26 +3,26 @@ package com.example.be.auth;
 import com.example.be.auth.dto.LoginRequest;
 import com.example.be.auth.dto.LoginResponse;
 import com.example.be.auth.dto.SignupRequest;
-import com.example.be.security.JwtProvider;
+import com.example.be.auth.service.AuthService;
+import com.example.be.auth.util.AuthHttpHelper;
+import com.example.be.auth.validator.AuthValidator;
 import com.example.be.security.JwtUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @Tag(name = "Auth", description = "인증 API")
@@ -32,7 +32,8 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
   private final AuthService authService;
-  private final JwtProvider jwtProvider;
+  private final AuthValidator authValidator;
+  private final AuthHttpHelper authHttpHelper;
   private final JwtUtils jwtUtils;
 
   @Operation(summary = "회원가입", description = "새로운 사용자를 등록합니다.")
@@ -53,13 +54,7 @@ public class AuthController {
 
     // refreshToken 분리해서 따로 HttpOnly에 저장
     String refreshToken = response.refreshToken();
-    ResponseCookie responseCookie = ResponseCookie.from("refreshToken", refreshToken)
-        .maxAge(3600)
-        .path("/api/auth/refresh")    // 해당 url이 요청할때만 전송
-        .secure(true)
-        .httpOnly(true)
-        .sameSite("Lax")
-        .build();
+    ResponseCookie responseCookie = authHttpHelper.getResponseCookie(refreshToken);
 
     response = new LoginResponse(response.accessToken(), null);
 
@@ -77,73 +72,46 @@ public class AuthController {
   public ResponseEntity<Void> logout(HttpServletRequest request) {
     String token = jwtUtils.resolveToken(request);
 
-    if (token == null || token.isBlank()) {
-      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-          .build();
-    }
-
     try {
-      jwtProvider.validateToken(token);
-    } catch (io.jsonwebtoken.JwtException e) {
+      token = authValidator.validateAndGetToken(token);
+    } catch (io.jsonwebtoken.JwtException | IllegalArgumentException e) {
       return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
           .build();
     }
 
     authService.logout(token);
-    ResponseCookie cookie = ResponseCookie.from("refreshToken", "")
-        .maxAge(0)
-        .path("/api/auth/refresh")
-        .httpOnly(true)
-        .build();
+    ResponseCookie cookie = authHttpHelper.createLogoutCookie();
 
     return ResponseEntity.ok()
+        .header(HttpHeaders.SET_COOKIE, cookie.toString())
         .build();
   }
 
   @PostMapping("/refresh")
   public ResponseEntity<LoginResponse> refresh(HttpServletRequest request) {
-    String refreshToken = extractRefreshToken(request);
-
-    if (refreshToken == null || refreshToken.isBlank()) {
-      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-          .build();
-    }
+    String refreshToken = authHttpHelper.extractRefreshToken(request);
 
     try {
-      jwtProvider.validateToken(refreshToken);
-    } catch (io.jsonwebtoken.JwtException e) {
+      refreshToken = authValidator.validateAndGetToken(refreshToken);
+    } catch (io.jsonwebtoken.JwtException | IllegalArgumentException e) {
       return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
           .build();
     }
 
-    LoginResponse response = authService.refresh(refreshToken);
-    ResponseCookie responseCookie = ResponseCookie.from("refreshToken", refreshToken)
-        .maxAge(3600)
-        .path("/api/auth/refresh")
-        .secure(true)
-        .httpOnly(true)
-        .sameSite("Lax")
-        .build();
+    LoginResponse response;
 
+    try {
+      response = authService.refresh(refreshToken);
+    } catch (BadCredentialsException e) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+          .build();
+    }
+
+    ResponseCookie responseCookie = authHttpHelper.getResponseCookie(refreshToken);
     response = new LoginResponse(response.accessToken(), null);
 
     return ResponseEntity.ok()
         .header(HttpHeaders.SET_COOKIE, responseCookie.toString())
         .body(response);
-  }
-
-  private String extractRefreshToken(HttpServletRequest request) {
-    // 1. request에서 header 분리
-    Cookie[] cookies = request.getCookies();
-    if (cookies == null) {
-      return null;
-    }
-
-    for (Cookie cookie : request.getCookies()) {
-      if ("refreshToken".equals(cookie.getName())) {
-        return cookie.getValue();
-      }
-    }
-    return null;
   }
 }
